connection string dialog also refactored out
